JSON number is float64, convert to int64.
I'm setting the planned_end_date to an int64 in timelord, but it's still
getting parsed out as a float64 here. Since I was converting the
interface{} to an int64, things were blowing up. The code down the line
needs an int64, so convert it immediately to hopefully prevent further
panics.

+++ main.go
@@ -587,7 +587,7 @@ func (a *AnalysesApp) UpdateByID(ctx context.Context) http.Handler {
 		}
 
 		if _, ok = jobpatch["planned_end_date"]; ok {
-			dbpatch["planned_end_date"] = strconv.FormatInt(jobpatch["planned_end_date"].(int64), 10)
+			dbpatch["planned_end_date"] = strconv.FormatInt(int64(jobpatch["planned_end_date"].(float64)), 10)
 		}
 
 		if _, ok = jobpatch["description"]; ok {
